test: Build fuzz targets into separate executables
Adaptation of btc@2ca632e5b44a8385989c8539cc4e30e60fdee16c
